+++ src/renderer/topBarItems.js
@@ -1,7 +1,7 @@
 import { useState, useCallback } from 'react';
 import propTypes from 'prop-types';
 import styled from '@emotion/styled';
-import { ROUTE_GLOBAL_PAGE } from '../routes';
+import { ROUTE_SYNC_VIEW } from '../routes';
 import { GlobalPageIcon } from './components/GlobalPage/GlobalPage';
 import { layout } from './components/styles';
 import * as strings from '../strings';
@@ -50,7 +50,7 @@ const TopBarItem = function ({ extension }) {
 
   return (
     <TopBarExtension
-      onClick={() => extension.navigate(ROUTE_GLOBAL_PAGE)}
+      onClick={() => extension.navigate(ROUTE_SYNC_VIEW)}
       onMouseEnter={handleMouseEnter}
       onMouseLeave={handleMouseLeave}
     >