+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## UNRELEASED
 
 - Changed timestamp format in `Created at` and `Synced at` fields in all the "View details" entity panels.
+- Updated Mirantis Container Cloud icon.
 
 ## v5.1.0